+++ fabric/src/main/java/io/github/flemmli97/debugutils/fabric/DebugUtilsFabric.java
@@ -1,6 +1,6 @@
 package io.github.flemmli97.debugutils.fabric;
 
-import io.github.flemmli97.debugutils.Commands;
+import io.github.flemmli97.debugutils.DebugCommands;
 import io.github.flemmli97.debugutils.DebugToggles;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
@@ -9,7 +9,7 @@ import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 public class DebugUtilsFabric implements ModInitializer {
     @Override
     public void onInitialize() {
-        CommandRegistrationCallback.EVENT.register(((dispatcher, reg, dedicated) -> Commands.register(dispatcher)));
-        ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> DebugToggles.onLogin(handler.getPlayer()));
+        CommandRegistrationCallback.EVENT.register(((dispatcher, reg, dedicated) -> DebugCommands.register(dispatcher)));
+        ServerPlayConnectionEvents.DISCONNECT.register((handler, server) -> DebugToggles.onLogout(handler.getPlayer()));
     }
 }